Add initil support for sensors
